drm/vc4: hdmi: Add a set_timings callback
commit  32f577fa07f22374d33d72b69cfd461d5948256f from
https://github.com/raspberrypi/linux.git rpi-5.8.y

Similarly to the previous patches, the timings setup in the HDMI controller
of the BCM2711 is slightly different, mostly because it supports higher
resolutions and thus needed more spaces for the various timings, resulting
in the register layout changing.

Let's add a callback for that as well.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>