drm/vc4: hdmi: Add a set_timings callback
commit  8665a332b6cd3b5e7b910421c722d58694c5f367 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

Similarly to the previous patches, the timings setup in the HDMI controller
of the BCM2711 is slightly different, mostly because it supports higher
resolutions and thus needed more spaces for the various timings, resulting
in the register layout changing.

Let's add a callback for that as well.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>