MLK-23221: drm/bridge: nwl-dsi: Call panel enable early
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Right now, if we have a panel attached (referenced by panel_bridge), it
is still treated as bridge in the display pipe-line. Since we need to
make sure that the DPI reset de-asserted after the panel is initialized,
but the callbacks pre_enable and enable from drm_bridge helper functions
are executed in reverse order, the best way is to manually call enable
on the panel_bridge right after our pre_enable call.

Without this manual call, the order of execution is as follows:
1. panel_bridge pre_enable (translated into panel prepare)
2. nwl-dsi pre_enable
3. nwl-dsi enable
4. panel_bridge enable (translated into panel enable)

Since there is no way to move 4. before 3., this is the reason for which
step 4. is manually executed right after step 2.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
(cherry picked from commit d8c1f57f1bc8fa32eb56976423239d2592c10c4a)
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>