ARM: dts: bcm283x: Unify CMA configuration
commit  0746dbb9c56000ddddaa474f2285c498d16a0f61 from
https://github.com/raspberrypi/linux.git rpi-5.4.y

commit c5a1e5375d19bd4001c59dc5d482ac5b1ba51cbf upstream.

With the introduction of the Raspberry Pi 4 we were forced to explicitly
configure CMA's location, since arm64 defaults it into the ZONE_DMA32
memory area, which is not good enough to perform DMA operations on that
device. To bypass this limitation a dedicated CMA DT node was created,
explicitly indicating the acceptable memory range and size.

That said, compatibility between boards is a must on the Raspberry Pi
ecosystem so this creates a common CMA DT node so as for DT overlays to
be able to update CMA's properties regardless of the board being used.

Signed-off-by: Nicolas Saenz Julienne <nsaenzjulienne@suse.de>
Reviewed-by: Phil Elwell <phil@raspberrypi.org>
Signed-off-by: Florian Fainelli <f.fainelli@gmail.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>