drm/vc4: crtc: Add function to compute FIFO level bits
commit  88ccf423d87aacf4317b1e9484c91d47f0cd8262 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

The longer FIFOs in vc5 pixelvalves means that the FIFO full level
doesn't fit in the original register field and that we also have a
secondary field. In order to prepare for this, let's move the registers
fill part to a helper function.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>