usb: xhci: Raspberry Pi FW loader for VIA VL805
commit  1df8e0d057984d13fd69970805df7c903cd2bb25 from
https://github.com/raspberrypi/linux.git rpi-5.6.y

The VL805 FW may either be loaded from an SPI EEPROM or alternatively
loaded directly by the VideoCore firmware. A PCI reset will reset
the VL805 XHCI controller on the Raspberry Pi4 requiring the firmware
to be reloaded if an SPI EEPROM is not present.

Use a VideoCore mailbox to trigger the loading of the VL805
firmware (if necessary) after a PCI reset.

Signed-off-by: Tim Gover <tim.gover@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>