usb: xhci: Raspberry Pi FW loader for VIA VL805
commit  73ac0d63cd820265ff53cc2ecf22d7534caaed01 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

The VL805 FW may either be loaded from an SPI EEPROM or alternatively
loaded directly by the VideoCore firmware. A PCI reset will reset
the VL805 XHCI controller on the Raspberry Pi4 requiring the firmware
to be reloaded if an SPI EEPROM is not present.

Use a VideoCore mailbox to trigger the loading of the VL805
firmware (if necessary) after a PCI reset.

Signed-off-by: Tim Gover <tim.gover@raspberrypi.org>
Signed-off-by: Meng Li <Meng.Li@windriver.com>