usb: host: xhci: add support for Light Reset for DWC3 DRD
commit  d90c34a4ecd6c919c7c3e4dc1f613e27ee736acb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

This patch adds support for Light Reset in Host Driver when
the core is in DRD mode.

There is a requirement of LRESET instead of RESET (Hard) in DWC3
controller when it is switching to the peripheral mode.

Signed-off-by: Piyush Mehta <piyush.mehta@xilinx.com>
Signed-off-by: Manish Narani <manish.narani@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>`
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>