reset: simple: Add reset callback
commit  ee0520467081dbc56a89e2bd9e2bf0fdb747f678 from
https://github.com/raspberrypi/linux.git rpi-5.8.y

The reset-simple code lacks a reset callback that is still pretty easy to
implement. The only real thing to consider is the delay needed for a device
to be reset, so let's expose that as part of the reset-simple driver data.

Cc: Philipp Zabel <p.zabel@pengutronix.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>