reset: simple: Add reset callback
commit  586b570ed04ea725f6df058db225245c95aab88d from
https://github.com/raspberrypi/linux.git rpi-5.6.y

The reset-simple code lacks a reset callback that is still pretty easy to
implement. The only real thing to consider is the delay needed for a device
to be reset, so let's expose that as part of the reset-simple driver data.

Cc: Philipp Zabel <p.zabel@pengutronix.de>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>