net: macb: Increment rx bd head after allocating skb and buffer
commit  e8afa2221ccb54d23ab0fe64cc2a67740fa59e78 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.4

In gem_rx_refill rx_prepared_head is incremented at the beginning of
the while loop prepared the skb and data buffers. If the skb or data
buffer allocation fails, this BD will be unusable BDs until the head
loops back to the same BD (and obviously buffer allocation succeeds).
In the unlikely event that there's a string of allocation failures,
there will be an equal number of unusable BDs and an inconsistent RX
BD chain. Hence increment the head at the end of the while loop to be
clean.

Signed-off-by: Harini Katakam <harini.katakam@xilinx.com>
Acked-by: Radhey Shyam Pandey <radhey.shyam.pandey@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>