staging: vchiq_arm: Give vchiq children DT nodes
commit  922229086913bf6850ee4448b29ed5e411bf801e from
https://github.com/raspberrypi/linux.git rpi-5.8.y

vchiq kernel clients are now instantiated as platform drivers rather
than using DT, but the children of the vchiq interface may still
benefit from access to DT properties. Give them the option of a
a sub-node of the vchiq parent for configuration and to allow
them to be disabled.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>