MLK-23367 usb: dwc3: drd: enter device role by default
commit  from
https://source.codeaurora.org/external/imx/linux-imx

Current role change via role switch class keeps the previous
role if the parter removed, with this, the typec port as source
will keeps host role after typec parter removed, this makes the
udc is not registered so can't enable gadget function until the
typec port becomes to be power sink(e.g. connects a power source
to the typec port), this is not convinent, also keeping the USB
port to be at device mode instead of host mode if the port is
empty, has benefit on power saving, so this patch change the usb
to be device mode whenever the target usb role is none, this is
also align with latest upstream patch.

Reviewed-by: Peter Chen <peter.chen@nxp.com>
Signed-off-by: Li Jun <jun.li@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>