octeontx2-pf: Fix the page reference and free race in the otx2_alloc_rbuf()
In the current implementation of otx2_alloc_rbuf(), we only increment
the page reference after all the spaces are allocated. Before that the
page reference is only 1 set by alloc_pages(). But the allocated buffers
may be freed before all the space are allocated, in that situation
the pages would be freed while there are still some buffers using it.
In order to fix this issue, we steal the theory of page_frag_alloc()
by adding a max enough number to the page reference at first, and
then subtract extra reference after all the spaces are allocated.

Signed-off-by: Kevin Hao <haokexin@gmail.com>